requirements.txt
azure_storage==0.37.0
langchain==0.3.18
langchain_chroma==0.2.1
langchain_community==0.3.17
langchain_ollama==0.2.3
langchain_text_splitters==0.3.6
protobuf==5.29.3
streamlit==1.41.1
streamlit_ext==0.1.15
torch==2.6.0+cpu
